Update network_blocker.py to work with multiple processes
