+++ network_blocker.py
@@ -8,7 +8,12 @@
 #
 import os
 import sys
+
 import socket
+
+import subprocess
+
+from threading import Thread
 
 from typing import List, Tuple
 
@@ -31,6 +36,7 @@ def scan_network(ip_range: str) -> List[Tuple[str, str, str]]:
     for _, rcv in answered:
         ip = rcv[ARP].psrc
         mac = rcv[Ether].src
+
         try:
             hostname = socket.gethostbyaddr(ip)[0]
         except socket.herror:
@@ -40,11 +46,23 @@ def scan_network(ip_range: str) -> List[Tuple[str, str, str]]:
     return device_list
 
 
-def generate_devices_table(device_list: List[Tuple[str, str, str]]):
-    headers = ["Index", "IP Address", "MAC Address", "Hostname"]
+def generate_devices_table(
+    device_list: List[Tuple[str, str, str]], blocked_devices, gateway_ip: str
+):
+    headers = ["Index", "IP Address", "MAC Address", "Hostname", "Blocked"]
     table = [
-        (index, ip, mac, hostname)
-        for index, (ip, mac, hostname) in enumerate(device_list)
+        [
+            index if device[0] != gateway_ip else "Gateway",
+            device[0],
+            device[1],
+            device[2],
+            "Yes"
+            if any(process.args[-2] == device[0] for process in blocked_devices)
+            else "No"
+            if device[0] != gateway_ip
+            else "N/A",
+        ]
+        for index, device in enumerate(device_list)
     ]
 
     return tabulate(table, headers=headers, tablefmt="grid")
@@ -54,23 +72,47 @@ def block_device(
     device: Tuple[str, str, str],
     gateway_ip: str,
     iface: str,
-):
+) -> subprocess.Popen:
     target_ip, target_mac, hostname = device
 
-    # send the arp spoofing packets
-    arpspoof_cmd = f"sudo arpspoof -i {iface} -t {target_ip} {gateway_ip}"
-    os.system(arpspoof_cmd)
+    # create the arp spoofing command
+    arpspoof_cmd = ["sudo", "arpspoof", "-i", iface, "-t", target_ip, gateway_ip]
+
+    # start the process without echoing the output
+    with open(os.devnull, "w") as devnull:
+        process = subprocess.Popen(
+            arpspoof_cmd, stdout=devnull, stderr=devnull, start_new_session=True
+        )
+    # return the process
+    return process
+
+
+def processTerminator(blocked_devices):
+    for process in blocked_devices:
+        process.terminate()
+        process.wait()
+    print("All processes terminated.")
 
 
 def print_menu():
     os.system("clear")  # clear the terminal
 
-    print("\n#                Welcome to Network Blocker              #")
-    print("         Script to block a device on your network          \n")
-    print("                                  Developed by FakuVenturi \n")
+    print("\n#                      Welcome to Network Blocker                    #")
+    print("               Script to block a device on your network                \n")
+    print("                                              Developed by FakuVenturi \n")
 
 
-if __name__ == "__main__":
+def print_complete_menu(device_list, blocked_devices, IPAddr, gateway_ip):
+    print_menu()
+
+    print(generate_devices_table(device_list, blocked_devices, gateway_ip), "\n")
+
+    print("Your IP Address is: ", IPAddr, "\n")
+
+    print("Blocked devices: ", len(blocked_devices), "\n")
+
+
+def main():
     print_menu()
 
     if os.geteuid() != 0:
@@ -79,11 +121,11 @@ if __name__ == "__main__":
 
     print("Scanning network...")
 
+    ip_range = "192.168.1.0/24"  # the ip range to scan
+
     iface = (
         os.popen("ip -o -4 route show to default | awk '{print $5}'").read().strip("\n")
     )  # get the interface name and remove the trailing newline
-
-    ip_range = "192.168.1.0/24"  # the ip range to scan
 
     gateway_ip = (
         os.popen("ip -o -4 route show to default | awk '{print $3}'").read().strip("\n")
@@ -98,71 +140,145 @@ if __name__ == "__main__":
     # scan the network and get the list of devices
     device_list = scan_network(ip_range)
 
+    blocked_devices = []
+
     if not device_list:
         print("No devices found.")
         sys.exit(1)
 
-    # generate the table of devices
-    devices_table = generate_devices_table(device_list)
+    print_complete_menu(device_list, blocked_devices, IPAddr, gateway_ip)
 
-    print_menu()
+    print("\n")
 
-    print(devices_table, "\n")
-
-    print("           Your IP Address is: ", IPAddr, "\n\n\n")
-
-    # loop to get the user's choice
-    while True:
-        try:
-            # get the user's choice
-            choice = int(input("Enter the index of the device or -1 to exit: "))
-
-            if choice == -1:
-                print("\nExiting...")
-                break
-            elif 0 <= choice < len(device_list):
-                # check if the user selected the gateway and prevent it
-                if device_list[choice][0] == gateway_ip:
-                    print_menu()
-
-                    print(devices_table, "\n")
-
-                    print("           Your IP Address is: ", IPAddr, "\n")
-
-                    print("             You cannot block your gateway.\n")
-                    continue
-
-                print("\nselected device: ", choice)
-                selected_device = device_list[choice]
+    try:
+        while True:
+            try:
+                # get the user's choice
                 print(
-                    "Blocking",
-                    selected_device[0],
+                    "  Enter the index of the device to block/unblock\n  0 to block all\n -3 to unblock all\n -2 to rescan the network\n -1 to exit (Ctrl+C)\n"
                 )
-                print("Press Ctrl+C and wait to stop blocking the device.\n")
+                choice = int(input("~> "))
 
-                block_device(selected_device, gateway_ip, iface)
+                if choice == -1:
+                    # exit the script
+                    break
+                elif choice == -2:
+                    # rescan the network
+                    os.system("clear")
+                    print_menu()
+                    print("Scanning network...")
+                    device_list = scan_network(ip_range)
+                    print_complete_menu(
+                        device_list, blocked_devices, IPAddr, gateway_ip
+                    )
+                elif choice == -3:
+                    # unblock all devices
+                    if blocked_devices:
+                        print_complete_menu(
+                            device_list, blocked_devices, IPAddr, gateway_ip
+                        )
+                        print("Unblocking all devices...")
+                        for process in blocked_devices:
+                            process.terminate()
+                            process.wait()
+                        blocked_devices.clear()
+                        print_complete_menu(
+                            device_list, blocked_devices, IPAddr, gateway_ip
+                        )
+                    else:
+                        print_complete_menu(
+                            device_list, blocked_devices, IPAddr, gateway_ip
+                        )
+                        print("No devices are blocked.\n")
+                elif choice == 0:
+                    # block all devices
+                    if len(blocked_devices) < len(device_list) - 1:
+                        print_complete_menu(
+                            device_list, blocked_devices, IPAddr, gateway_ip
+                        )
+                        print("Blocking all devices...")
+                        for device in device_list:
+                            # check if the device is the gateway and prevent it
+                            if device[0] == gateway_ip:
+                                continue
+                            blocked_devices.append(
+                                block_device(device, gateway_ip, iface)
+                            )
+                        print_complete_menu(
+                            device_list, blocked_devices, IPAddr, gateway_ip
+                        )
+                    else:
+                        print_complete_menu(
+                            device_list, blocked_devices, IPAddr, gateway_ip
+                        )
+                        print("All devices are already blocked.\n")
+                elif 0 < choice < len(device_list):
+                    # block/unblock the selected device
+                    # check if the user selected the gateway and prevent it
+                    if device_list[choice][0] == gateway_ip:
+                        print_complete_menu(
+                            device_list, blocked_devices, IPAddr, gateway_ip
+                        )
 
-                print_menu()
+                        print("You cannot block your gateway.\n")
+                        continue
 
-                print(devices_table, "\n")
+                    print_complete_menu(
+                        device_list, blocked_devices, IPAddr, gateway_ip
+                    )
+                    print("selected device: ", choice)
+                    selected_device = device_list[choice]
 
-                print("           Your IP Address is: ", IPAddr, "\n")
-            else:
-                print_menu()
+                    if any(
+                        process.args[-2] == selected_device[0]
+                        for process in blocked_devices
+                    ):
+                        print("\nUnblocking", selected_device[0])
 
-                print(devices_table, "\n")
+                        for process in blocked_devices:
+                            if process.args[-2] == selected_device[0]:
+                                process.terminate()
+                                process.wait()
+                                blocked_devices.remove(process)
+                    else:
+                        print("\nBlocking", selected_device[0])
 
-                print("           Your IP Address is: ", IPAddr, "\n")
+                        process = block_device(selected_device, gateway_ip, iface)
 
-                print("           Invalid choice. Please try again.\n")
-        except ValueError:
-            print_menu()
+                        blocked_devices.append(process)
 
-            print(devices_table, "\n")
+                    print_complete_menu(
+                        device_list, blocked_devices, IPAddr, gateway_ip
+                    )
+                else:
+                    print_complete_menu(
+                        device_list, blocked_devices, IPAddr, gateway_ip
+                    )
 
-            print("           Your IP Address is: ", IPAddr, "\n")
+                    print("Invalid choice. Please try again.\n")
+            except ValueError:
+                print_complete_menu(device_list, blocked_devices, IPAddr, gateway_ip)
 
-            print("          Invalid input. Please enter a number.    \n")
-        except KeyboardInterrupt:
-            print("\nExiting...")
-            break
+                print("Invalid input. Please enter a number.    \n")
+    except KeyboardInterrupt:
+        if blocked_devices:
+            print_complete_menu(device_list, blocked_devices, IPAddr, gateway_ip)
+            print("Terminating all processes...\n")
+            t = Thread(target=processTerminator, args=(blocked_devices,))
+            t.start()
+            t.join()
+
+    finally:
+        if blocked_devices:
+            print_complete_menu(device_list, blocked_devices, IPAddr, gateway_ip)
+            print("Terminating all processes...\n")
+            t = Thread(target=processTerminator, args=(blocked_devices,))
+            t.start()
+            t.join()
+
+
+if __name__ == "__main__":
+    try:
+        main()
+    except KeyboardInterrupt:
+        print("\nClosing...")
